frontend/pyproject.toml
[tool.poetry]
name = "aiogram-soloway-bot"
version = "0.1.0"
description = ""
authors = ["German Golcov <gerleffx2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aerich = "0.6.1"
aiogram = "2.17"
aiohttp = "3.8.1"
aioredis = "2.0.0"
aiosignal = "1.2.0"
aiosqlite = "0.16.1"
async-timeout = "4.0.1"
asyncpg = "0.25.0"
attrs = "20.3.0"
Babel = "2.9.1"
bcrypt = "3.2.0"
beautifulsoup4 = "4.9.3"
bs4 = "0.0.1"
certifi = "2021.10.8"
cffi = "1.15.0"
chardet = "3.0.4"
charset-normalizer = "2.0.9"
click = "8.0.3"
cryptography = "3.3.2"
ddlparse = "1.10.0"
dictdiffer = "0.9.0"
docutils = "0.18.1"
emoji = "1.6.1"
fake-useragent = "0.1.11"
frozenlist = "1.2.0"
gunicorn = "20.0.4"
hiredis = "1.1.0"
idna = "2.10"
iniconfig = "1.1.1"
iso8601 = "0.1.16"
Jinja2 = "3.0.3"
loguru = "0.5.3"
lxml = "4.6.4"
MarkupSafe = "2.0.1"
multidict = "4.7.6"
packaging = "21.3"
pluggy = "0.13.1"
py = "1.9.0"
pycparser = "2.21"
pydantic = "1.8.2"
pyparsing = "3.0.6"
pypika-tortoise = "0.1.1"
pytest = "6.1.2"
pytz = "2020.5"
requests = "2.25.1"
six = "1.15.0"
soupsieve = "2.0.1"
toml = "0.10.2"
tomlkit = "0.7.2"
tortoise-orm = "0.17.8"
typing_extensions = "4.0.1"
urllib3 = "1.26.7"
yarl = "1.5.1"
backoff = "^2.0.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "loader.TORTOISE_CONFIG"
location = "./db_models/migrations"
src_folder = "./."